backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML Libraries
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
transformers==4.37.0
torch==2.1.2
sentence-transformers==2.3.1

# Speech-to-Text
openai-whisper==20231117
yt-dlp==2024.1.19
pydub==0.25.1
SpeechRecognition==3.10.1

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.2
sumy==0.11.0

# Vector Database
pinecone-client==3.0.2
chromadb==0.4.22

# YouTube Integration
google-api-python-client==2.116.0
youtube-transcript-api==0.6.2

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.26.0
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
